More adjustments to get the imports working again.
