+++ game_system/__init__.py
@@ -3,9 +3,9 @@
 ### IMPORTS ###
 from game_system.character import Character
 
-from game_system.item import Item, ItemFactory
+from game_system.item_factory import ItemFactory
 
-from game_system.equipment import Equipment, EquipmentFactory
+from game_system.equipment_factory import EquipmentFactory
 
 from game_system.exceptions import CharacterInvalidStatTypeException
 from game_system.exceptions import GameSystemMismatchException
@@ -15,6 +15,9 @@ from game_system.exceptions import ItemNotInInventoryException
 from game_system.exceptions import NotCharacterException
 
 from game_system.charfile import CharacterFile
+
+import game_system.none
+import game_system.shadowrun
 
 ### GLOBALS ###
