+++ game_system/none/__init__.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python3
+
+### IMPORTS ###
+#from game_system.shadowrun.character import ShadowRunCharacter
+
+from game_system.none.item import Item
+
+from game_system.none.equipment import Equipment
+
+### GLOBALS ###
+
+### FUNCTIONS ###
+
+### CLASSES ###
