+++ game_system/shadowrun/__init__.py
@@ -3,9 +3,9 @@
 ### IMPORTS ###
 from game_system.shadowrun.character import ShadowRunCharacter
 
-from game_system.shadowrun.item import ShadowRunItem, ShadowRunItemFactory
+from game_system.shadowrun.item import ShadowRunItem
 
-from game_system.shadowrun.equipment import ShadowRunEquipment, ShadowRunEquipmentFactory
+from game_system.shadowrun.equipment import ShadowRunEquipment
 
 ### GLOBALS ###
